backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
openai
numpy
faiss-cpu
pandas
python-dotenv